other:暂存

+++ migrations/20240703094709-create-cart.js
@@ -18,7 +18,7 @@ module.exports = {
         allowNull: false
       },
       totalPrice: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false
       },
       quantity: {
